+++ src/App.jsx
@@ -1,33 +1,38 @@
-import React from "react";
-// import { Counter } from "./features/counter/Counter";
+import React, { useState, createContext } from "react";
 import { Routes, Route } from "react-router-dom";
 import "./App.css";
-import { Login } from "./view/login";
-import { Dashboard } from "./view/dashboard/index";
-import { Income } from "./view/income/index";
-import { Expense } from "./view/expense/index";
-import { Register } from "./view/register/index";
-import {Home} from './components/home/'
-import { NotFound } from "./view/notfound";
-import { IncomeDetails } from './view/incomedetails/index';
-import { ExpenseDetails } from "./view/expensedetails";
+import { Login } from "./view/login/login";
+import { Dashboard } from "./view/dashboard/dashboard";
+import { Income } from "./view/income/income";
+import { Expense } from "./view/expense/expense";
+import { Register } from "./view/register/register";
+import { Navigate } from "./components/navigate/navigate";
+import { NotFound } from "./view/notfound/notfound";
+import { IncomeDetails } from "./view/incomedetails/incomedetails";
+import { ExpenseDetails } from "./view/expensedetails/expensedetails";
+import { Protected } from "./utils/utils";
 
+const AuthContext = createContext(null);
+console.log(AuthContext);
 function App() {
+  const [token, setToken] = useState(null);
+
   return (
-    <div className="App">
+    <AuthContext.Provider value={token}>
+      <Navigate />
       <Routes>
-        <Route path="/" element={<Home />}>
+        <Route path="/">
           <Route index element={<Login />} />
           <Route path="register" element={<Register />} />
           <Route path="dashboard" element={<Dashboard />} />
           <Route path="income" element={<Income />} />
-          <Route path="income/:incomeId" element={< IncomeDetails/>} />
+          <Route path="income/:incomeId" element={<IncomeDetails />} />
           <Route path="expense" element={<Expense />} />
           <Route path="expense/:expenseId" element={<ExpenseDetails />} />
           <Route path="*" element={<NotFound />} />
         </Route>
       </Routes>
-    </div>
+    </AuthContext.Provider>
   );
 }
